Pass audio callback to interrupt handler
With this change the callback gets passed to the interrupt handler every
time it's needed. This makes the API more flexible. For example, it is
possible to keep audio handler in a global context, where its parameters
can be changed, as it can be always passed as closure's context.

This also simplifies the audio example to simply pass through audio.

Signed-off-by: Petr Horacek <hrck@protonmail.com>