Set correct format on Seed 1.1 and Patch SM
Before this patch, these two were using I2S format. While that was not
a problem for a signal passthrough, the signal was not properly
reproduced after processing.

With this patch, the audio example does a simple processing to expose
this problem and the two codecs are setting format to 24-bit LJ.

Signed-off-by: Petr Horacek <hrck@protonmail.com>